+++ src/views/login/components/ForgetPasswordForm/index.vue
@@ -1,7 +1,7 @@
 <template>
-    <van-form v-if="getShow" ref="formRef" class="forget-form" @submit="handleReset">
+    <VanForm v-if="getShow" ref="formRef" class="forget-form" @submit="handleReset">
         <!-- 通用字段：由 data.ts 的 forgetPasswordFields 驱动 -->
-        <van-field
+        <VanField
             v-for="field in forgetPasswordFields"
             :key="field.key"
             v-model="(formData as Record<string, string>)[field.key]"
@@ -16,13 +16,13 @@
                 <i :class="[field.icon, 'forget-form__icon']" />
             </template>
             <template v-if="field.hasButton" #button>
-                <van-button size="small" type="primary" style="border-radius: var(--ds-radius-sm)">
+                <VanButton size="small" type="primary" style="border-radius: var(--ds-radius-sm)">
                     {{ field.buttonText }}
-                </van-button>
+                </VanButton>
             </template>
-        </van-field>
+        </VanField>
 
-        <van-button
+        <VanButton
             class="forget-form__btn forget-form__btn--primary"
             type="primary"
             block
@@ -30,12 +30,12 @@
             :loading="loading"
         >
             重置密码
-        </van-button>
+        </VanButton>
 
-        <van-button class="forget-form__btn forget-form__btn--secondary" block @click="handleBackLogin">
+        <VanButton class="forget-form__btn forget-form__btn--secondary" block @click="handleBackLogin">
             返回登录
-        </van-button>
-    </van-form>
+        </VanButton>
+    </VanForm>
 </template>
 
 <script setup lang="ts">